refactor: move cli to ckb-bin
