+++ ckb-bin/src/tests/bats_tests/ckb_run_replay.bats
