+++ ckb-bin/src/tests/bats_tests/cli.bats
