+++ ckb-bin/src/tests/bats_tests/graceful_shutdown.bats
