+++ ckb-bin/src/tests/bats_tests/init_reset.bats
