+++ ckb-bin/src/tests/bats_tests/load_notify_config.bats
