+++ ckb-bin/src/tests/bats_tests/peer_id.bats
