+++ ckb-bin/src/tests/cli.rs
