+++ util/app-config/Cargo.toml
@@ -9,9 +9,7 @@ homepage = "https://github.com/nervosnetwork/ckb"
 repository = "https://github.com/nervosnetwork/ckb"
 
 [dependencies]
-clap = { version = "=4.4", features = ["string", "wrap_help"] }
 serde = { version = "1.0", features = ["derive"] }
-serde_plain = "0.3.0"
 serde_json = "1.0"
 toml = "0.5"
 path-clean = "0.1.0"
@@ -24,7 +22,6 @@ ckb-pow = { path = "../../pow", version = "= 0.119.0-pre" }
 ckb-resource = { path = "../../resource", version = "= 0.119.0-pre" }
 ckb-build-info = { path = "../build-info", version = "= 0.119.0-pre" }
 ckb-types = { path = "../types", version = "= 0.119.0-pre" }
-ckb-constant = { path = "../constant", version = "= 0.119.0-pre" }
 secio = { version = "0.6", package = "tentacle-secio" }
 multiaddr = { version = "0.3.0", package = "tentacle-multiaddr" }
 rand = "0.8"
