+++ util/app-config/src/configs/mod.rs
@@ -28,4 +28,4 @@ pub use rpc::{Config as RpcConfig, Module as RpcModule};
 pub use store::Config as StoreConfig;
 pub use tx_pool::{BlockAssemblerConfig, TxPoolConfig};
 
-pub(crate) use network::{generate_random_key, read_secret_key, write_secret_to_file};
+pub use network::{generate_random_key, read_secret_key, write_secret_to_file};
