+++ util/app-config/src/tests/mod.rs
@@ -1,3 +1,2 @@
 mod app_config;
-mod cli;
 mod legacy;
